requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
pillow==10.1.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-mock==3.11.1